requirements.txt
blinker==1.4
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
Flask==2.2.2
Flask-Mail==0.9.1
Flask-WTF==1.0.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.0
python-http-client==3.3.7
sendgrid==6.9.7
starkbank-ecdsa==2.1.0
Werkzeug==2.2.2
WTForms==3.0.1
